refactor(cli): connectors commands move to src/cli/connectors.ts
Pure move of the `connectors` command group (list, install, remove,
doctor) out of bin/lcm.ts's main() into a new registerConnectorsCommands
export, following the shape established for memory and bench.
